Completely changed the way the routes work.
Now you can specify routes in two ways, with text files (similar to yaml)
and php.

Still needs a lot of work, specially dealing with errors and not
found routes (404)

But this version is a working proof of concept.

Got rid of yaml files because parsing is easier this way and
we don't need an external dependency for parsing yaml files.

I created a new .route file type